backend/pyproject.toml
[project]
name = "deeplearn-week5-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.21",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "starlette>=0.41.0",
    "itsdangerous>=2.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
